+++ app/src/main/AndroidManifest.xml
@@ -2,27 +2,19 @@
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools">
     <!-- 인터넷 권한 선언(volley 라이브러리를 사용하기 위한 설정 -->
-    <uses-permission android:name="android.permission.INTERNET" />
-
-    <!--   GPS 관련  -->
-    <permission android:name="wonpa.alwaysweb.com.myloctest.permission.MAPS_RECEIVE"
-        android:protectionLevel="signature" />
-
-    <!-- Google Map -->
+    <uses-permission android:name="android.permission.INTERNET" /> <!-- GPS 관련 -->
+    <permission
+        android:name="wonpa.alwaysweb.com.myloctest.permission.MAPS_RECEIVE"
+        android:protectionLevel="signature" /> <!-- Google Map -->
     <uses-permission android:name="wonpa.alwaysweb.com.myloctest.permission.MAPS_RECEIVE" />
-    <uses-permission android:name="com.google.android.providers.gsf.permission.READ_GSERVICES" />
-    <!--  인공위성 GPS 연동   -->
-    <uses-permission android:name="android.permission.ACCESS_FINE_LOCATION" />
-    <!--  기지국을 통한 GPS 연동  -->
-    <uses-permission android:name="android.permission.ACCESS_COARSE_LOCATION"/>
-
-
-    <!-- 걸음수 측정   -->
-    <uses-permission android:name="android.permission.ACTIVITY_RECOGNITION"/>
+    <uses-permission android:name="com.google.android.providers.gsf.permission.READ_GSERVICES" /> <!-- 인공위성 GPS 연동 -->
+    <uses-permission android:name="android.permission.ACCESS_FINE_LOCATION" /> <!-- 기지국을 통한 GPS 연동 -->
+    <uses-permission android:name="android.permission.ACCESS_COARSE_LOCATION" /> <!-- 걸음수 측정 -->
+    <uses-permission android:name="android.permission.ACTIVITY_RECOGNITION" />
 
     <uses-feature
         android:glEsVersion="0x00020000"
-        android:required="true"/>
+        android:required="true" />
 
     <application
         android:allowBackup="true"
@@ -32,12 +24,11 @@
         android:label="@string/app_name"
         android:supportsRtl="true"
         android:theme="@style/Theme.Move_Whole_Project"
-        tools:targetApi="31"
-        android:usesCleartextTraffic="true">
-        <!-- usesCleartextTraffic이 true여야 http 통신이 가능하다        -->
+        android:usesCleartextTraffic="true"
+        tools:targetApi="31">
 
-        <!-- Google Map 관련 라이브러리 사용 설정   -->
-
+        <!-- usesCleartextTraffic이 true여야 http 통신이 가능하다 -->
+        <!-- Google Map 관련 라이브러리 사용 설정 -->
         <uses-library android:name="com.google.android.maps" />
         <uses-library
             android:name="org.apache.http.legacy"
@@ -46,26 +37,28 @@
         <meta-data
             android:name="com.google.android.maps.v2.API_KEY"
             android:value="AIzaSyCXkQaWmvaMN5ANiRdNncfuyPuHSM01v44" />
-
         <meta-data
             android:name="com.google.android.gms.version"
-            android:value="@integer/google_play_services_version" />
+            android:value="@integer/google_play_services_version" /> <!-- GPS 기능 액티비티 -->
+        <activity
+            android:name=".Main_GPS.Activity_GPS"
+            android:exported="false" /> <!-- 로그인 회원가입 기능 액티비티 -->
+        <activity
+            android:name=".Main_GPS.Activity_Main"
+            android:exported="false" /> <!-- 하단 메뉴 액티비티-->
 
-        <!--  GPS 기능 액티비티     -->
-        <activity android:name=".Main_GPS.Activity_Main"
-            android:exported="false"/>
-
-        <!-- 로그인 회원가입 기능 액티비티       -->
-        <activity android:name=".Register_Login.Activity_Login"
-            android:exported="false"/>
-        <activity android:name=".Register_Login.Activity_Signup"
-            android:exported="false"/>
-        <activity android:name=".Register_Login.Activity_SignupDone"
-            android:exported="false"/>
-        <activity android:name=".Register_Login.YearMonthPicker"
-            android:exported="false"/>
-
-        <!-- 시작화면 액티비티     -->
+        <activity
+            android:name=".Register_Login.Activity_Login"
+            android:exported="false" />
+        <activity
+            android:name=".Register_Login.Activity_Signup"
+            android:exported="false" />
+        <activity
+            android:name=".Register_Login.Activity_SignupDone"
+            android:exported="false" />
+        <activity
+            android:name=".Register_Login.YearMonthPicker"
+            android:exported="false" /> <!-- 시작화면 액티비티 -->
         <activity
             android:name=".Register_Login.Activity_Start"
             android:exported="true">
@@ -75,7 +68,6 @@
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-
     </application>
 
 </manifest>
