+++ app/src/main/java/com/example/move_whole_project/Fragment/RecordFragment.java
@@ -1,0 +1,21 @@
+package com.example.move_whole_project.Fragment;
+
+import android.os.Bundle;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import androidx.fragment.app.Fragment;
+
+import com.example.move_whole_project.R;
+
+public class RecordFragment extends Fragment {
+
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup container,
+                             Bundle savedInstanceState) {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_record, container, false);
+    }
+
+}
