+++ app/src/main/res/drawable/ic_mission.xml
@@ -1,0 +1,7 @@
+<vector android:autoMirrored="true" android:height="24dp"
+    android:viewportHeight="496" android:viewportWidth="496"
+    android:width="24dp" xmlns:android="http://schemas.android.com/apk/res/android">
+    <path android:fillColor="#000000" android:pathData="M416,248V116.69L299.31,0H0v496h416v-96h80V248H416zM304,27.31L388.69,112H304V27.31zM400,480H16V16h272v112h112v120h-16v16h-5.34l-28.35,-21.26L328,198.11V144h-45.42L264,190.46V216l5.75,24h-84.3l-9.64,48.19l10.53,42.12L160,352.6V336H32v128h128v-90.45l30.66,-25.94l1.58,6.33L211.06,392h152.26l8,-8H384v16h16V480zM280,215.02v-21.48L293.42,160H312v41.89l25.69,51.38L373.34,280H384v88h-19.31l-8,8H220.94l-13.4,-26.79L192.19,287.8l6.36,-31.8h91.7L280,215.02zM75.72,432h15.21L144,387.1V448H48v-96h96v14.14L85.07,416H84.28l-13.62,-20.44l-13.31,8.87L75.72,432zM480,384h-80V264h80V384z"/>
+    <path android:fillColor="#000000" android:pathData="M160,69.55l37.17,-31.45L186.84,25.89L160,48.6V32H32v128h128V69.55zM144,62.14L85.07,112H84.28L70.66,91.56l-13.31,8.87L75.72,128h15.21L144,83.1V144H48V48h96V62.14z"/>
+    <path android:fillColor="#000000" android:pathData="M160,221.55l37.17,-31.45l-10.33,-12.22L160,200.6V184H32v128h128V221.55zM144,214.14L85.07,264H84.28l-13.62,-20.44l-13.31,8.87L75.72,280h15.21L144,235.1V296H48v-96h96V214.14z"/>
+</vector>
