+++ app/src/main/res/drawable/ic_record.xml
@@ -1,0 +1,13 @@
+<vector android:autoMirrored="true" android:height="24dp"
+    android:viewportHeight="502" android:viewportWidth="502"
+    android:width="24dp" xmlns:android="http://schemas.android.com/apk/res/android">
+    <path android:fillColor="#000000" android:pathData="M492,152.5H65.15l153.54,-91.99C220.45,76.79 234.27,89.5 251,89.5c16.77,0 30.6,-12.76 32.32,-29.08l65.56,39.1c4.74,2.83 10.88,1.27 13.71,-3.47c2.83,-4.74 1.28,-10.88 -3.47,-13.71L256.12,20.91c-3.16,-1.88 -7.1,-1.88 -10.26,0.01L26.23,152.5H10c-5.52,0 -10,4.48 -10,10v310c0,5.52 4.48,10 10,10h482c5.52,0 10,-4.48 10,-10v-310C502,156.98 497.52,152.5 492,152.5zM251,44.5c6.89,0 12.5,5.61 12.5,12.5s-5.61,12.5 -12.5,12.5s-12.5,-5.61 -12.5,-12.5S244.11,44.5 251,44.5zM482,462.5H20v-290h462V462.5z"/>
+    <path android:fillColor="#000000" android:pathData="M389.88,123.97l20,11.93c1.61,0.96 3.37,1.41 5.11,1.41c3.41,0 6.73,-1.74 8.6,-4.88c2.83,-4.74 1.28,-10.88 -3.47,-13.71l-20,-11.93c-4.74,-2.83 -10.88,-1.28 -13.71,3.47C383.58,115.01 385.14,121.14 389.88,123.97z"/>
+    <path android:fillColor="#000000" android:pathData="M107,336.06c-5.52,0 -10,4.48 -10,10V409.5c0,5.52 4.48,10 10,10c5.52,0 10,-4.48 10,-10v-63.44C117,340.54 112.52,336.06 107,336.06z"/>
+    <path android:fillColor="#000000" android:pathData="M203,305.48c-5.52,0 -10,4.48 -10,10V409.5c0,5.52 4.48,10 10,10c5.52,0 10,-4.48 10,-10v-94.02C213,309.96 208.52,305.48 203,305.48z"/>
+    <path android:fillColor="#000000" android:pathData="M299,262.68c-5.52,0 -10,4.48 -10,10v14.57c0,5.52 4.48,10 10,10c5.52,0 10,-4.48 10,-10v-14.57C309,267.16 304.52,262.68 299,262.68z"/>
+    <path android:fillColor="#000000" android:pathData="M299,305.11c-5.52,0 -10,4.48 -10,10V409.5c0,5.52 4.48,10 10,10c5.52,0 10,-4.48 10,-10v-94.39C309,309.59 304.52,305.11 299,305.11z"/>
+    <path android:fillColor="#000000" android:pathData="M395,305.48c-5.52,0 -10,4.48 -10,10V409.5c0,5.52 4.48,10 10,10c5.52,0 10,-4.48 10,-10v-94.02C405,309.96 400.52,305.48 395,305.48z"/>
+    <path android:fillColor="#000000" android:pathData="M370.26,271.72l19.04,10.26c1.51,0.81 3.13,1.2 4.74,1.2c3.56,0 7.01,-1.9 8.81,-5.26c2.62,-4.86 0.8,-10.93 -4.06,-13.55l-19.04,-10.26c-4.86,-2.62 -10.93,-0.8 -13.55,4.06C363.58,263.03 365.39,269.1 370.26,271.72z"/>
+    <path android:fillColor="#000000" android:pathData="M107,311.46c1.07,0 2.16,-0.17 3.23,-0.54l94.08,-32.1c0.38,-0.13 0.76,-0.28 1.13,-0.46l93.32,-45.16l39.5,21.29c4.86,2.62 10.93,0.8 13.55,-4.06c2.62,-4.86 0.8,-10.93 -4.06,-13.55l-44,-23.71c-2.83,-1.52 -6.21,-1.6 -9.1,-0.2l-97.37,47.13l-93.5,31.91c-5.23,1.78 -8.02,7.47 -6.23,12.69C98.95,308.84 102.84,311.46 107,311.46z"/>
+</vector>
