+++ app/src/main/res/layout/activity_gps.xml
@@ -1,0 +1,73 @@
+<?xml version="1.0" encoding="utf-8"?>
+<LinearLayout
+    xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:tools="http://schemas.android.com/tools"
+    android:layout_width="match_parent"
+    android:layout_height="match_parent"
+    android:orientation="vertical"
+    tools:context=".Main_GPS.Activity_GPS">
+
+    <Button
+        android:id="@+id/coordinate_value"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:text="좌표값 현재위치"/>
+
+    <!-- 위도, 경도 표시-->
+    <TextView
+        android:id="@+id/value_text"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"/>
+
+    <!-- 변환된 주소 출력 -->
+    <TextView
+        android:id="@+id/txt"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:textSize="16dp"/>
+
+    <!-- 걸음 수 -->
+    <TextView
+        android:id="@+id/steps_count"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="걸음 수 : "
+        android:textSize="16dp"/>
+
+    <!-- 이동 거리 -->
+    <TextView
+        android:id="@+id/distances"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:text="이동 거리 : "
+        android:textSize="16dp"/>
+
+    <!-- 날짜 -->
+    <TextView
+        android:id="@+id/current_date"
+        android:layout_width="wrap_content"
+        android:layout_height="wrap_content"
+        android:textSize="16dp"/>
+
+
+    <!-- 현재 내 위치 보기-->
+    <Button
+        android:id="@+id/my_location"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:text="현재 내 위치"/>
+
+    <!-- 걸음 수 초기화-->
+    <Button
+        android:id="@+id/resetButton"
+        android:layout_width="match_parent"
+        android:layout_height="wrap_content"
+        android:text="걸음수 초기화"/>
+
+    <fragment
+        android:id="@+id/map"
+        class="com.google.android.gms.maps.SupportMapFragment"
+        android:layout_width="match_parent"
+        android:layout_height="403dp" />
+
+</LinearLayout>
