+++ app/src/main/res/menu/bottom_menu.xml
@@ -1,0 +1,30 @@
+<?xml version="1.0" encoding="utf-8"?>
+<menu xmlns:android="http://schemas.android.com/apk/res/android"
+    xmlns:app="http://schemas.android.com/apk/res-auto">
+
+    <item
+        android:id="@+id/home"
+        android:enabled="true"
+        android:icon="@drawable/ic_home"
+        android:title="홈"
+        app:showAsAction="always"
+        />
+
+    <item
+        android:id="@+id/mission"
+        android:enabled="true"
+        android:icon="@drawable/ic_mission"
+        android:title="미션"
+        app:showAsAction="always"
+        />
+
+
+    <item
+        android:id="@+id/record"
+        android:enabled="true"
+        android:icon="@drawable/ic_record"
+        android:title="기록"
+        app:showAsAction="always"
+        />
+
+</menu>
